pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ad-neuro-diagnostics"
version = "0.1.0"
description = "Brain-inspired creative diagnostics for short video ads using cached TRIBE artifacts."
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name = "Ivan Lambev"}]
dependencies = [
  "typer>=0.16,<1.0",
  "pandas>=2.2,<3.0",
  "numpy>=2.0,<3.0",
  "matplotlib>=3.9,<4.0",
  "mne>=1.8,<2.0",
  "nibabel>=5.3,<6.0",
  "nilearn>=0.11,<1.0",
  "scikit-learn>=1.5,<2.0",
  "tabulate>=0.9,<1.0",
  "pillow>=10.4,<12.0",
  "moviepy>=2.2.1,<3.0",
  "soundfile>=0.13,<1.0",
  "pyyaml>=6.0,<7.0",
]

[project.optional-dependencies]
test = [
  "pytest>=8.3,<9.0",
]
download = [
  "yt-dlp>=2024.8.6",
]

[project.scripts]
ads = "ad_neuro_diagnostics.cli:app"

[tool.setuptools.packages.find]
include = ["ad_neuro_diagnostics*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
